requirements.txt
numpy==1.24.4
pillow==10.4.0
pycryptodome==3.22.0
pycryptodomex==3.22.0
pytest==7.4.0
cryptography==40.0.2